update 文字表示機能を分割
振る舞い的に、ファイルとディレクトリが違うのでis-aが成立しないと考え、継承関係も解消
